Refactor: split multi-test classes to separate files
